[in-memory] bail on bad navigate() method, re-org tests
